+++ leetcode/hash_table/two_sum.cpp
@@ -8,7 +8,7 @@ class Solution {
  public:
   std::vector<int> twoSum(const std::vector<int>& nums, int target) {
     std::map<int, int> table;
-    for (auto i = 0; i < nums.size(); ++i) {
+    for (auto i = 0; i < static_cast<int>(nums.size()); ++i) {
       auto value = target - nums[i];
       if (table.find(value) != table.end()) {
         return {table[value], i};
